Use asterisk as emphasis marker
After years of using Markdown, with an underscore for emphasis
and an asterisk for importance (strong), I’m now switching to
using asterisks for everything: lists, importance, and emphasis.

The reason for this is that asterisks are better: they typically
don’t occur in normal text, whereas dashes, pluses, and underscores
do (in programming related topics at least).  Therefore they are
objectively less likely to cause accidental markup.